% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_map.R
\name{bird_map}
\alias{bird_map}
\title{Interactive Map of Bird Detections}
\usage{
bird_map(data, title = "Bird Detections")
}
\arguments{
\item{data}{A data frame containing birdnet data, sensor.id, lat, and
long columns.}

\item{title}{Character. A title for the legend}
}
\value{
A Leaflet interactive map with bird detections by sensor.id
}
\description{
Interactive Map of Bird Detections
}
