% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_list.R
\name{birdnet_list}
\alias{birdnet_list}
\title{List of species detected by BirdNET}
\usage{
birdnet_list(df, sort = "n.days")
}
\arguments{
\item{df}{A data frame with the detections}

\item{sort}{Character. Indicates how the list should be sorted out. Use
"n.days" for number of days detected, "n.calls" for number of calls detected,
or "call.rate" for call rate.}
}
\value{
A data frame.
}
\description{
List of species detected by BirdNET
}
\examples{
\dontrun{
birdListSundarban <- birdnetList(BirdnetDetections)
}
}
