% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdnetVisualizeR-package.R
\docType{package}
\name{BirdnetVisualizeR-package}
\alias{BirdnetVisualizeR}
\alias{BirdnetVisualizeR-package}
\title{BirdnetVisualizeR: Fast plots for BirdNET results}
\description{
This package provides tools to create interactive and static plots for BirdNET data.
}
\author{
\strong{Maintainer}: Francisco Rivas \email{fcorivasf@gmail.com}

}
\keyword{internal}
