% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_chart.R
\name{pheno_chart}
\alias{pheno_chart}
\title{Phenological charts}
\usage{
pheno_chart(df, plot_title = "", min_days = 5, sort = "start", desc = TRUE)
}
\arguments{
\item{df}{A data frame obtained with "import_birdnet" function.}

\item{plot_title}{Character. A title for the plot (use quotation marks).}

\item{min_days}{Numeric. Minimum number of days of detection for a species
to be included in the plot.}

\item{sort}{Character. Specifies how to order the species list vertically.
Options:
"start" (default) sorts by the first day of detection;
"n.days" sorts by the total number of days that the species was detected;
"n.calls" sorts by the total number of calls detected by species;
"call.rate" sorts by the calling rate of each species (n.calls/n.days)
"common.name" sorts by the common name of the species, in English.
"scientific.name" sorts by the binomial scientific name of the species.}

\item{desc}{Logical. Whether to sort in descending order.}
}
\value{
a Plotly object.
}
\description{
This function creates a "phenochart" with all the species
detected with BirdNET. The chart is a stacked tile plot with a heatmap
indicating the number of calls per species per day, and a separate tile plot
at the bottom indicating the total number of species detected each day.
Francisco Rivas, April 2024
}
\examples{
\dontrun{
pheno_chart(BirdnetDetections,
             plot_title = "BirdNET - Sundarban Mangrove Forest S4A17644",
             sort = "start", min_days = 15)
# filter some species
selectedsP <- BirdnetDetections |>
 filter(Common.name \%in\% c("Mangrove Pitta", "Indian White-eye",
                           "Mangrove Whistler","Common Tailorbird"))
pheno_chart(selectedsP,  plot_title = "BirdNET - Sundarban Mangrove Forest S4A17644")


selectedsP <- BirdnetDetections |>
 filter(Common.name \%in\% c("Large-tailed Nightjar","Oriental Scops-Owl" ))

pheno_chart(selectedsP,  plot_title = "BirdNET - Sundarban Mangrove Forest S4A17644")


}
}
