% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenochart_single.R
\name{phenochart_single}
\alias{phenochart_single}
\title{Phenological Chart - Single Species}
\usage{
phenochart_single(df, sp, aggregate_by = "week", add_sensor_id = FALSE)
}
\arguments{
\item{df}{A data frame}

\item{sp}{Character. The species' binomial (scientific name).}

\item{aggregate_by}{Character. Level of data aggregataion. Otions are "day"
and "week".}

\item{add_sensor_id}{Logical. Whether to add the sensor ID to the plot.}
}
\value{
A heatmap  plot.
}
\description{
Phenological Chart - Single Species
}
\examples{
\dontrun{
phenochart_singls(data, "Troglodytes aedon",
                  aggregate_by = "week",
                  add_sensor_id = TRUE)
}
}
