% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_plots.R
\name{taxa_plots}
\alias{taxa_plots}
\title{Create Faceted Barplots for Taxonomic Summary Results}
\usage{
taxa_plots(summary_list, n_top = 15, fill_color = "steelblue", base_size = 12)
}
\arguments{
\item{summary_list}{A list of taxonomic summary tables (output from taxa_summary functions)}

\item{n_top}{Number of top items to display in each plot. Use NULL to show all.
Default is 15.}

\item{fill_color}{Color for the bars. Default is "steelblue".}

\item{base_size}{Base font size for the plots. Default is 12.}
}
\value{
A list of ggplot objects, one for each summary table
}
\description{
This function generates faceted barplots for each table in the taxonomic
summary results list, with options to limit the number of bars displayed.
}
