% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_summary.R
\name{taxon_summary}
\alias{taxon_summary}
\title{Taxonomic summary}
\usage{
taxon_summary(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A tibble with number of species, families, and orders.
}
\description{
Calculates the number of species, families, and orders of birds
represented in a data frame
}
\examples{
\dontrun{
taxon_summary(my_data)
}
}
