% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_treemap.R
\name{taxon_treemap}
\alias{taxon_treemap}
\title{Treemap of bird call detections organized by order and species}
\usage{
taxon_treemap(data, save_png = TRUE, prefix = "")
}
\arguments{
\item{data}{a data frame containing BirdNET detections and taxonomic
information, including "order" and "species" columns.}

\item{save_png}{Logical. Whether to save plot as png.}

\item{prefix}{Characer. Prefix for file name.}
}
\value{
a treemap plot
}
\description{
Treemap of bird call detections organized by order and species
}
