% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_sp_plot.R
\name{top_sp_plot}
\alias{top_sp_plot}
\title{Top N species}
\usage{
top_sp_plot(birdnet_list, site_id)
}
\arguments{
\item{birdnet_list}{a data frame obtained with the "birdnet_list" function.}

\item{site_id}{A string (text) with an identifier for the site, sensor, or
plot (use quotations)}
}
\value{
a bar plot depicting the 10 most detected species in a site.
}
\description{
Get a stacked bar plot with top 10 species (i.e., the most detected)
}
\examples{
\dontrun{
top_sp_plot(S4A17644_list, " - sensor S4A17644")
}
}
