% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocal_hyperdominance.R
\name{vocal_hyperdominance}
\alias{vocal_hyperdominance}
\title{Vocal hyperdominance per grouping variable}
\usage{
vocal_hyperdominance(
  data,
  group_var = species,
  past_half = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame with BirdNET outputs and, optionally,
taxonomic data.}

\item{group_var}{Character. The name of the column to be used as
grouping variable (e.g., species, Common.name, order, family)}

\item{past_half}{Logical. Whether to show the species that pushes
past 50%}
}
\value{
A tibble
}
\description{
Vocal hyperdominance per grouping variable
}
