% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_species.R
\name{top_species}
\alias{top_species}
\title{Top species summary}
\usage{
top_species(data, site_id = "", n_species = 10)
}
\arguments{
\item{data}{a data frame containing BirdNET detections.}

\item{site_id}{A string (text) with an identifier for the site, sensor, or
plot (use quotations)}

\item{n_species}{Number of top species to summarize.Default is 10}
}
\value{
a bar plot depicting the 10 most detected species in a site.
}
\description{
Get a stacked bar plot with top 10 species (i.e., the most detected)
}
\examples{
\dontrun{
top_species(S4A17644_list, " - sensor S4A17644")
}
}
